se modifica nombres de las clases y se eliminan repetidas
