+++ src/vista/CalculoGanancias.java
@@ -7,15 +7,17 @@ import javax.swing.JFrame;
 import javax.swing.JPanel;
 import javax.swing.border.EmptyBorder;
 import javax.swing.JButton;
-import javax.swing.JTextField;
+import javax.swing.JLabel;
 import javax.swing.JTable;
+import javax.swing.JTextField;
 import javax.swing.table.DefaultTableModel;
 
-public class Vista_Almacen extends JFrame {
+public class CalculoGanancias extends JFrame {
 
 	private JPanel contentPane;
-	private JTextField textField;
 	private JTable table;
+	private JTextField tfd_Desde;
+	private JTextField tfd_hasta;
 
 	/**
 	 * Launch the application.
@@ -24,7 +26,7 @@ public class Vista_Almacen extends JFrame {
 		EventQueue.invokeLater(new Runnable() {
 			public void run() {
 				try {
-					Vista_Almacen frame = new Vista_Almacen();
+					CalculoGanancias frame = new CalculoGanancias();
 					frame.setVisible(true);
 				} catch (Exception e) {
 					e.printStackTrace();
@@ -36,31 +38,18 @@ public class Vista_Almacen extends JFrame {
 	/**
 	 * Create the frame.
 	 */
-	public Vista_Almacen() {
-		setTitle("Almac\u00E9n");
+	public CalculoGanancias() {
 		setResizable(false);
+		setTitle("Calculo de Ganancias");
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		setBounds(100, 100, 473, 351);
+		setBounds(100, 100, 450, 300);
 		contentPane = new JPanel();
 		contentPane.setBorder(new EmptyBorder(5, 5, 5, 5));
 		setContentPane(contentPane);
 		contentPane.setLayout(null);
 		
-		JButton btn_todos_los_Productos = new JButton("Todos los productos");
-		btn_todos_los_Productos.setBounds(10, 11, 142, 23);
-		contentPane.add(btn_todos_los_Productos);
-		
-		textField = new JTextField();
-		textField.setBounds(10, 45, 177, 20);
-		contentPane.add(textField);
-		textField.setColumns(10);
-		
-		JButton btn_Buscar = new JButton("Buscar");
-		btn_Buscar.setBounds(193, 44, 89, 23);
-		contentPane.add(btn_Buscar);
-		
 		JButton btn_Volver = new JButton("Volver");
-		btn_Volver.setBounds(378, 11, 69, 23);
+		btn_Volver.setBounds(348, 11, 76, 23);
 		contentPane.add(btn_Volver);
 		
 		table = new JTable();
@@ -88,27 +77,41 @@ public class Vista_Almacen extends JFrame {
 				{null, null, null, null},
 			},
 			new String[] {
-				"Productos", "Actual", "M\u00EDnimo", "M\u00E1ximo"
+				"N\u00BA", "Fecha", "Hora", "Ganancias"
 			}
 		));
-		table.setBounds(10, 76, 437, 141);
+		table.setBounds(185, 51, 228, 148);
 		contentPane.add(table);
 		
-		JButton btn_Realizar_Lista_Compra = new JButton("Realizar lista de la compra");
-		btn_Realizar_Lista_Compra.setBounds(10, 236, 157, 23);
-		contentPane.add(btn_Realizar_Lista_Compra);
+		JLabel lbl_Desde = new JLabel("Desde");
+		lbl_Desde.setBounds(10, 64, 46, 14);
+		contentPane.add(lbl_Desde);
 		
-		JButton btn_Ver_Listas_Compras = new JButton("Ver listas de la compra");
-		btn_Ver_Listas_Compras.setBounds(178, 236, 142, 23);
-		contentPane.add(btn_Ver_Listas_Compras);
+		JLabel lbl_Hasta = new JLabel("Hasta");
+		lbl_Hasta.setBounds(10, 95, 46, 14);
+		contentPane.add(lbl_Hasta);
 		
-		JButton btn_Modificar = new JButton("Modificar");
-		btn_Modificar.setBounds(368, 228, 79, 23);
-		contentPane.add(btn_Modificar);
+		tfd_Desde = new JTextField();
+		tfd_Desde.setBounds(48, 61, 86, 20);
+		contentPane.add(tfd_Desde);
+		tfd_Desde.setColumns(10);
 		
-		JButton btn_Nuevo = new JButton("Nuevo");
-		btn_Nuevo.setBounds(368, 263, 79, 23);
-		contentPane.add(btn_Nuevo);
+		tfd_hasta = new JTextField();
+		tfd_hasta.setBounds(48, 92, 86, 20);
+		contentPane.add(tfd_hasta);
+		tfd_hasta.setColumns(10);
+		
+		JLabel lbl_Total_Ganancias = new JLabel("Total Ganancias");
+		lbl_Total_Ganancias.setBounds(30, 161, 86, 14);
+		contentPane.add(lbl_Total_Ganancias);
+		
+		JLabel lbl_Ganancias = new JLabel("284,76");
+		lbl_Ganancias.setBounds(40, 186, 46, 14);
+		contentPane.add(lbl_Ganancias);
+		
+		JButton btn_Imprimir = new JButton("Imprimir");
+		btn_Imprimir.setBounds(185, 210, 76, 23);
+		contentPane.add(btn_Imprimir);
 	}
 
 }
